// Haly.WebApp/ClientApp/generated/haly/models/DuplicateType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Haly API
 * An ASP.NET Core Web API that adds quality of life improvements to Spotify and helps with music exploration
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DuplicateType = {
    None: 'None',
    Some: 'Some',
    All: 'All'
} as const;
export type DuplicateType = typeof DuplicateType[keyof typeof DuplicateType];


export function DuplicateTypeFromJSON(json: any): DuplicateType {
    return DuplicateTypeFromJSONTyped(json, false);
}

export function DuplicateTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): DuplicateType {
    return json as DuplicateType;
}

export function DuplicateTypeToJSON(value?: DuplicateType | null): any {
    return value as any;
}